+++ src/Umbraco.Cms.Api.Management/Controllers/Element/RecycleBin/DeleteElementFolderRecycleBinController.cs
@@ -1,28 +1,36 @@
 using Asp.Versioning;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Umbraco.Cms.Api.Management.Factories;
 using Umbraco.Cms.Core;
+using Umbraco.Cms.Core.Actions;
 using Umbraco.Cms.Core.Models;
 using Umbraco.Cms.Core.Security;
+using Umbraco.Cms.Core.Security.Authorization;
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Core.Services.OperationStatus;
+using Umbraco.Cms.Web.Common.Authorization;
+using Umbraco.Extensions;
 
 namespace Umbraco.Cms.Api.Management.Controllers.Element.RecycleBin;
 
 [ApiVersion("1.0")]
 public class DeleteElementFolderRecycleBinController : ElementRecycleBinControllerBase
 {
+    private readonly IAuthorizationService _authorizationService;
     private readonly IBackOfficeSecurityAccessor _backOfficeSecurityAccessor;
     private readonly IElementContainerService _elementContainerService;
 
     public DeleteElementFolderRecycleBinController(
+        IAuthorizationService authorizationService,
         IEntityService entityService,
         IElementPresentationFactory elementPresentationFactory,
         IBackOfficeSecurityAccessor backOfficeSecurityAccessor,
         IElementContainerService elementContainerService)
         : base(entityService, elementPresentationFactory)
     {
+        _authorizationService = authorizationService;
         _backOfficeSecurityAccessor = backOfficeSecurityAccessor;
         _elementContainerService = elementContainerService;
     }
@@ -34,6 +42,16 @@ public class DeleteElementFolderRecycleBinController : ElementRecycleBinControll
     [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status404NotFound)]
     public async Task<IActionResult> Delete(CancellationToken cancellationToken, Guid id)
     {
+        AuthorizationResult authorizationResult = await _authorizationService.AuthorizeResourceAsync(
+            User,
+            ElementPermissionResource.RecycleBin(ActionElementDelete.ActionLetter),
+            AuthorizationPolicies.ElementPermissionByResource);
+
+        if (!authorizationResult.Succeeded)
+        {
+            return Forbidden();
+        }
+
         Attempt<EntityContainer?, EntityContainerOperationStatus> result = await _elementContainerService.DeleteFromRecycleBinAsync(id, CurrentUserKey(_backOfficeSecurityAccessor));
 
         return result.Success
